Adds revamped prices logic and UX
It is now possible to define multiple base prices as well as temporary limited seasonal prices per ticket/package. The first base price is thereby not removable and is valid for ever and since ever (it replaces the old basic price). On top of this the user can define new base prices that are valid from a certain point in time.

Seasonal prices are limited in start time _and_ in end time. They also include year information, so a ticket/package can have different seasonal prices each year.