Removes deactivated accommodations, addons, packages, tickets and trips from view
Also disables all restoring. This basically makes the app fullfill the idea behind deactivated models: the application still needs them to function, but the user should not care about deactivated vs. deletion. It should be the same from the user's point of view.

Additionally, `parent_id` columns are added to the four tables, so models that are created while updating can refer to their parent model (for later analytic purposes).