Finish move to DigitalOcean live server with the Krystal server being the test server
This makes the laravel system to only be accessible from the /api folder inside the /dashboard folder. After adjusting some path in laravel files, this is not a problem. We therefore don't have to deal with Cross-Domain-Access controls, since everything runs under the same rms. subdomain (or rms-test).

Changes all routes in laravel to omit the api prefix.
Moves Krystal DB configuration in its own environment config.
Splits the .htaccess in half and removes the need for the exluded-routes block.
Adds a URL building (and redirecting) part into the dashboard/index.php file at the top.

All-in-all it's a good streamlining and was the obvious solution in hindsight ;)