Fix price calculation for a booking for when the booking is reserved, expired or confirmed
This new methods respects the date of when a ticket was added to the booking to calculate its price. This means that when a ticket is added to a (reserved|expired|confirmed) booking, but the price of the tickets has changed in the meantime, the old price is still calculated for the existing tickets while the new price is calculated for the new (added) ticket.

[skip ci]