Fix calculated price for accommodations
The price was calculated as an average per night, which was in turn not considered in all the calculations that are based on this price (and it cannot be requested that the other code should consider it, as it would incur unreasonable amounts of extra work). The price is now the overall price of the accommodation from check-in to check-out. The accommodation model gains a new attribute that contains the average price per night.