Trigger form submit when clicking the "Next" button on the Add Booking tab's Extra Info screen
Resolves #150, references #157

Previously, only the "Save" button at the bottom of the Extra Info form saved the values, but the "Next" button that the user is trained to click did NOT save them. Now, the "Next" button has been changed to first submit the form when on the Extra Info screen and then, after success, load the next tab.

#157 needs checking if it is also resolved by this fix.