Reduce number of server requests in Addons tab by 50%
This commit adds the created or updated model to the return data of the create or update request. This way the front-end does not have to request an updated model list "manually" anymore after the changing request has finished.

The commit also greatly simplifies and standardises the rendering and update logic and reduces required code. The simplification prepares the code structure for easy replacement by a future data-binding library, such as Ractive.js.