Fix bug in Manage Bookings where the reserved-until date was not shown correctly in the tooltips
The `reserved` attribute of the booking model was unfavourably named and as such overwritten by a boolean that marked a booking's state. The DB columns has been renamed to `reserved_until`, which better reflects that it contains the expiry date of a reservation, and all occurrences in the code should have been updated.