Fix My Mailing Lists tab layout
The issue was with the missuse of bootstrap classes, nameley the use of `.form-group` instead of the correct `.form-row`. This lead to the shifting of elements due to implied styles of the `.form-group`. Replacing those classes also allowed me to get rid of the extra `fieldset` that enclosed the former `.form-group` divs. This makes the HTML cleaner now.

I also changed the intendation of *3 tabs* (WTF?) back to one tab. We have a coding style here, and since ALL files in the whole system are made with 1 tab, all new files should also respect this setting. (Same with spaces instead of tabs in the `mailing-lists/js/script.js` file!)